Changed config.xml
After more reading, I found this in the PhoneGap Build documentation for
config.xml:
<feature>: the feature element can be used to specify which features
your application is using. If you specify features of the PhoneGap API,
those will be expanded to the appropriate Android and Windows Phone
permissions for you application.
So I added the <feature> tag for accessing the camera, hopefully that
will fix permissions for android.